pyproject.toml
[tool.poetry]
name = "timeclock"
version = "0.2.5"
description = "A simple timeclock bot for Discord"
authors = ["DLCHAMP <36091350+dlchamp@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^2.0.16"
disnake = "^2.8.1"
python-dotenv = "^1.0.0"
aiosqlite = "^0.19.0"
coloredlogs = "^15.0.1"
thefuzz = "^0.19.0"
python-levenshtein = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
